backend/requirements.txt
# LiveKit dependencies
livekit-server-sdk==0.6.0
livekit-api==0.4.0
livekit-agents==0.9.0

# Web framework
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0

# LLM integrations
openai==1.3.0
groq==0.4.0
httpx==0.25.0

# Audio processing
wave==0.0.2
pydub==0.25.1
speech-recognition==3.10.0

# Optional Twilio integration
twilio==8.10.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
typing-extensions==4.8.0
asyncio==3.4.3